Refactoring + CommonConfig class created
