+++ src/main/java/hu/drumbun/config/CommonConfig.java
@@ -1,0 +1,14 @@
+package hu.drumbun.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class CommonConfig {
+
+    @Bean
+    ModelMapper getModelMapper() {
+        return new ModelMapper();
+    }
+}
