+++ src/main/java/hu/drumbun/controller/user/UserController.java
@@ -1,7 +1,7 @@
-package hu.drumbun.controller;
+package hu.drumbun.controller.user;
 
 import hu.drumbun.entities.User;
-import hu.drumbun.service.UserService;
+import hu.drumbun.service.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
@@ -14,36 +14,36 @@ public class UserController {
     @Autowired
     UserService userService;
 
-    @RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE,value = "/users")
-    public List<User> getAllUsers(){
+    @RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE, value = "/users")
+    public List<User> getAllUsers() {
         return userService.getAllUsers();
     }
 
     @RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE, value = "/users/{id}")
-    public User getUserById(@PathVariable long id){
+    public User getUserById(@PathVariable long id) {
         return userService.getUserById(id);
     }
 
     @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, value = "/adduser")
-    public String addUser(@RequestBody User user){
-        userService.addUser(new User(user.getEmail(),user.getPassword()));
+    public String addUser(@RequestBody User user) {
+        userService.addUser(new User(user.getEmail(), user.getPassword()));
         return "User added to DB!";
     }
 
     @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, value = "/addfuser")
-    public String addForeignUser(@RequestBody User user){
-        userService.addUser(new User(user.getEmail(),user.getPassword(),user.getOauth_provider(),user.getOauth_uid()));
+    public String addForeignUser(@RequestBody User user) {
+        userService.addUser(new User(user.getEmail(), user.getPassword(), user.getOauth_provider(), user.getOauth_uid()));
         return "User added to DB!";
     }
 
     @RequestMapping(method = RequestMethod.DELETE, value = "/delete/{id}")
-    public String deleteUserById(@PathVariable long id){
+    public String deleteUserById(@PathVariable long id) {
         userService.removeUserById(id);
         return "User deleted from DB";
     }
 
     @RequestMapping(method = RequestMethod.GET, value = "/users/{email}")
-    public User getUserByEmail(@PathVariable String email){
+    public User getUserByEmail(@PathVariable String email) {
         return userService.getUserByEmail(email);
     }
 }
