+++ src/main/java/hu/drumbun/service/user/UserService.java
@@ -1,4 +1,4 @@
-package hu.drumbun.service;
+package hu.drumbun.service.user;
 
 import hu.drumbun.entities.User;
